Fix most remaining MSVC warnings (essentially loss of precision)
git-svn-id: https://svn.osgeo.org/gdal/trunk@31687 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/apps/gdalserver.c
@@ -159,7 +159,7 @@ static CPL_SOCKET CreateSocketAndBindAndListen(const char* pszService,
         if (nListenSocket == INVALID_SOCKET)
             continue;
 
-        if (bind(nListenSocket, psResultsIter->ai_addr,
+        if ((int)bind(nListenSocket, psResultsIter->ai_addr,
                  psResultsIter->ai_addrlen) != SOCKET_ERROR)
         {
             if( pnFamily )   *pnFamily =   psResultsIter->ai_family;
